+++ cli/dinghy/constants.rb
@@ -3,10 +3,15 @@ require 'fileutils'
 
 MEM_DEFAULT=2048
 CPU_DEFAULT=1
+DISK_DEFAULT=20_000
 
-BREW = Pathname.new(`brew --prefix`.strip)
-DINGHY = BREW+"opt/dinghy/etc"
-VAGRANT = BREW+"var/dinghy/vagrant"
-HOST_IP = "192.168.42.1"
-VM_IP = "192.168.42.10"
-HOME = ENV.fetch("HOME")
+# makes local dev easier
+if $0 == "bin/dinghy" || $0 =~ /rspec/
+  BREW = '/usr/local'
+  DINGHY = Pathname.new(File.expand_path("../..", File.dirname(__FILE__)))
+else
+  BREW = Pathname.new(`brew --prefix`.strip)
+  DINGHY = BREW+"opt/dinghy/etc"
+end
+HOME = Pathname.new(ENV.fetch("HOME"))
+HOME_DINGHY = HOME+'.dinghy'