+++ cli/dinghy/http_proxy.rb
@@ -1,22 +1,29 @@
 require 'stringio'
 
-require 'dinghy/vagrant'
+require 'dinghy/machine'
 
-# we ssh in, so this succeeds even if the env vars haven't been set yet
 class HttpProxy
   CONTAINER_NAME = "dinghy_http_proxy"
 
+  attr_reader :machine
+
+  def initialize(machine)
+    @machine = machine
+  end
+
   def up
     puts "Starting the HTTP proxy"
-    capture_output do
-      Vagrant.new.ssh("docker rm -fv #{CONTAINER_NAME}") rescue nil
+    System.capture_output do
+      docker.system("rm", "-fv", CONTAINER_NAME)
     end
-    Vagrant.new.ssh("docker run -d -p 80:80 -v /var/run/docker.sock:/tmp/docker.sock --name #{CONTAINER_NAME} codekitchen/dinghy-http-proxy")
+    docker.system("run", "-d", "-p", "80:80", "-v", "/var/run/docker.sock:/tmp/docker.sock", "--name", CONTAINER_NAME, "codekitchen/dinghy-http-proxy")
   end
 
   def status
-    output, _ = capture_output do
-      Vagrant.new.ssh("docker inspect -f '{{ .State.Running }}' #{CONTAINER_NAME}")
+    return "not running" if !machine.running?
+    
+    output, _ = System.capture_output do
+      docker.system("inspect", "-f", "{{ .State.Running }}", CONTAINER_NAME)
     end
 
     if output.strip == "true"
@@ -24,21 +31,11 @@ class HttpProxy
     else
       "not running"
     end
-  rescue # ehhhhhh
-    "not running"
   end
 
   private
 
-  def capture_output
-    prev_stdout = $stdout.dup
-    prev_stderr = $stderr.dup
-    $stdout.reopen(Tempfile.new("stdout"))
-    $stderr.reopen(Tempfile.new("stderr"))
-    yield
-    return $stdout.tap(&:rewind).read, $stderr.tap(&:rewind).read
-  ensure
-    $stdout.reopen(prev_stdout)
-    $stderr.reopen(prev_stderr)
+  def docker
+    @docker ||= Docker.new(machine)
   end
 end